Set up movie details fragment
